% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Settings.R
\name{CensoringTime}
\alias{CensoringTime}
\title{Censoring time (linear)}
\usage{
CensoringTime(N, p, Covariates, censoring_par, censoring_intercept)
}
\arguments{
\item{N}{Sample size.}

\item{p}{Number of covariates.}

\item{Covariates}{Covariate matrix.}

\item{censoring_par}{Covariate effect vector on censoring rate.}

\item{censoring_intercept}{Baseline (intercept) censoring rate.}
}
\value{
Censoring-time vector.
}
\description{
Censoring time (linear)
}
