% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Settings.R
\name{CensoringTime_nonlinear}
\alias{CensoringTime_nonlinear}
\title{Censoring time (nonlinear)}
\usage{
CensoringTime_nonlinear(N, p, Covariates)
}
\arguments{
\item{N}{Sample size.}

\item{p}{Number of covariates.}

\item{Covariates}{Covariate matrix.}
}
\value{
Censoring-time vector.
}
\description{
Censoring time (nonlinear)
}
