% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic.R
\name{DataFitting}
\alias{DataFitting}
\alias{DataFitting.ModelPar.ITT}
\alias{DataFitting.ModelPar.remove}
\alias{DataFitting.ModelPar.recensor}
\alias{DataFitting.ModelPar.TimeVar}
\alias{DataFitting.ModelPar.DRIVE.joint}
\alias{DataFitting.ModelPar.DRIVE.ML}
\title{Fit a treatment-switching model}
\usage{
DataFitting(ModelPar)

\method{DataFitting}{ModelPar.ITT}(ModelPar)

\method{DataFitting}{ModelPar.remove}(ModelPar)

\method{DataFitting}{ModelPar.recensor}(ModelPar)

\method{DataFitting}{ModelPar.TimeVar}(ModelPar)

\method{DataFitting}{ModelPar.DRIVE.joint}(ModelPar)

\method{DataFitting}{ModelPar.DRIVE.ML}(ModelPar)
}
\arguments{
\item{ModelPar}{A `ModelPar` object from [New_ModelPar()].}
}
\value{
A list with at least `Coef` and `Var`.
}
\description{
Generic that dispatches on the `ModelPar.<method>` class. Supported methods
are `ITT`, `remove`, `recensor`, `TimeVar`, `DRIVE.joint` and `DRIVE.ML`.
See [DRIVE_ML] for the user-supplied learner interface.
}
