% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Settings.R
\name{InitAssignment}
\alias{InitAssignment}
\title{Assign treatment via a linear propensity score}
\usage{
InitAssignment(N, p, Covariates, gamma)
}
\arguments{
\item{N}{Sample size.}

\item{p}{Number of covariates.}

\item{Covariates}{Covariate matrix.}

\item{gamma}{Coefficient vector for the propensity score linear predictor.}
}
\value{
Binary treatment vector.
}
\description{
Assign treatment via a linear propensity score
}
