% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Settings.R
\name{InitAssignment_Nonlinear}
\alias{InitAssignment_Nonlinear}
\title{Assign treatment via a nonlinear propensity score}
\usage{
InitAssignment_Nonlinear(N, p, Covariates)
}
\arguments{
\item{N}{Sample size.}

\item{p}{Number of covariates.}

\item{Covariates}{Covariate matrix.}
}
\value{
Binary treatment vector.
}
\description{
Assign treatment via a nonlinear propensity score
}
