% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Settings.R
\name{InitCovariates}
\alias{InitCovariates}
\title{Initialise measured (and optionally unmeasured) covariates}
\usage{
InitCovariates(N, p, p_U, unmeasured_Confounding)
}
\arguments{
\item{N}{Sample size.}

\item{p}{Number of measured covariates.}

\item{p_U}{Number of unmeasured confounders.}

\item{unmeasured_Confounding}{Function generating unmeasured confounders.}
}
\value{
Covariate matrix.
}
\description{
Initialise measured (and optionally unmeasured) covariates
}
