% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassBuilder.R
\name{New_SimuArg}
\alias{New_SimuArg}
\title{Construct a simulation-argument object}
\usage{
New_SimuArg(
  nrep,
  N,
  p,
  p_U,
  Scenario,
  max_t,
  theta,
  unmeasured_Confounding = function(N, p_U, ...) NULL,
  InitCovariates = function(N, p, ...) NULL,
  InitAssignment = function(N, p, Covariates, ...) NULL,
  SurvTime = function(N, p, Covariates, W, theta, ...) NULL,
  SwitchingTime = function(N, p, Covariates, Z, ...) NULL,
  CensoringTime = function(N, p, Covariates, ...) NULL,
  TransferX = function(N, p, Covariates, ...) NULL,
  Control = list(json_save = FALSE, save_path = "", Annotation = ""),
  ...
)
}
\arguments{
\item{nrep}{Number of replicates.}

\item{N}{Sample size.}

\item{p}{Number of measured covariates.}

\item{p_U}{Number of unmeasured confounders.}

\item{Scenario}{`"i"` (exogenous switching) or `"ii"` (endogenous switching).
The former names `"exogenous"` and `"endogenous"` remain accepted aliases.}

\item{max_t}{Administrative censoring time.}

\item{theta}{True treatment effect.}

\item{unmeasured_Confounding, InitCovariates, InitAssignment, SurvTime, SwitchingTime, CensoringTime, TransferX}{Data-generating building-block functions.}

\item{Control}{List of generation controls (e.g. `json_save`, `save_path`, `Annotation`).}

\item{...}{Additional scenario parameters (e.g. `gamma`, `alpha`, `beta`).}
}
\value{
An object of class `paste0("SimuArg.", Scenario)`.
}
\description{
Builds a `SimuArg.<Scenario>` object bundling the data-generating functions
and parameters used by [DataGenerating()].
}
