% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Settings.R
\name{SurvTime}
\alias{SurvTime}
\title{Survival time under a linear structural model}
\usage{
SurvTime(N, p, Covariates, W, theta, alpha, Z)
}
\arguments{
\item{N}{Sample size.}

\item{p}{Number of covariates.}

\item{Covariates}{Covariate matrix.}

\item{W}{Switching-time vector.}

\item{theta}{True treatment effect.}

\item{alpha}{Covariate effect vector.}

\item{Z}{Treatment assignment vector.}
}
\value{
List with `T_D` (observed) and `T_0` (control) event times.
}
\description{
Survival time under a linear structural model
}
