% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Settings.R
\name{SurvTime_endogenous}
\alias{SurvTime_endogenous}
\title{Survival time under a linear model with endogenous switching}
\usage{
SurvTime_endogenous(N, p, Covariates, W, theta, alpha, Z, T)
}
\arguments{
\item{N}{Sample size.}

\item{p}{Number of covariates.}

\item{Covariates}{Covariate matrix.}

\item{W}{Switching-time vector.}

\item{theta}{True treatment effect.}

\item{alpha}{Covariate effect vector.}

\item{Z}{Treatment assignment vector.}

\item{T}{Pre-drawn baseline event time.}
}
\value{
List with `T_D` and `T_0`.
}
\description{
Survival time under a linear model with endogenous switching
}
