% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Settings.R
\name{SurvTime_Nonlinear}
\alias{SurvTime_Nonlinear}
\title{Survival time under a nonlinear structural model}
\usage{
SurvTime_Nonlinear(N, p, Covariates, W, theta, Z)
}
\arguments{
\item{N}{Sample size.}

\item{p}{Number of covariates (must be 2).}

\item{Covariates}{Covariate matrix.}

\item{W}{Switching-time vector.}

\item{theta}{True treatment effect.}

\item{Z}{Treatment assignment vector.}
}
\value{
List with `T_D` and `T_0`.
}
\description{
Survival time under a nonlinear structural model
}
