% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Settings.R
\name{SwitchingTime}
\alias{SwitchingTime}
\title{Treatment-switching time (exogenous)}
\usage{
SwitchingTime(N, p, Covariates, Z, beta, diffcoef)
}
\arguments{
\item{N}{Sample size.}

\item{p}{Number of covariates.}

\item{Covariates}{Covariate matrix.}

\item{Z}{Treatment assignment vector.}

\item{beta}{Covariate effect vector on switching rate.}

\item{diffcoef}{Treatment-arm differential for switching rate.}
}
\value{
Switching-time vector.
}
\description{
Treatment-switching time (exogenous)
}
