% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Settings.R
\name{SwitchingTime_endogenous}
\alias{SwitchingTime_endogenous}
\title{Treatment-switching time (endogenous)}
\usage{
SwitchingTime_endogenous(N, p, Covariates, Z, beta, diffcoef, T, alpha)
}
\arguments{
\item{N}{Sample size.}

\item{p}{Number of covariates.}

\item{Covariates}{Covariate matrix.}

\item{Z}{Treatment assignment vector.}

\item{beta}{Covariate effect vector on switching rate.}

\item{diffcoef}{Treatment-arm differential for switching rate.}

\item{T}{Pre-drawn baseline event time.}

\item{alpha}{Covariate effect vector on the baseline hazard.}
}
\value{
Switching-time vector.
}
\description{
Treatment-switching time (endogenous)
}
