% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Settings.R
\name{unmeasured_Confounding}
\alias{unmeasured_Confounding}
\title{Generate unmeasured confounders (independent)}
\usage{
unmeasured_Confounding(N, p_U)
}
\arguments{
\item{N}{Sample size.}

\item{p_U}{Number of unmeasured confounders.}
}
\value{
An `N x p_U` matrix of uniform draws.
}
\description{
Generate unmeasured confounders (independent)
}
