% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Settings.R
\name{unmeasured_Confounding_dependent}
\alias{unmeasured_Confounding_dependent}
\title{Generate unmeasured confounders dependent on a measured covariate}
\usage{
unmeasured_Confounding_dependent(N, p, p_U, Covariates)
}
\arguments{
\item{N}{Sample size.}

\item{p}{Covariate index (column of `Covariates` used as the base).}

\item{p_U}{Number of unmeasured confounders.}

\item{Covariates}{Measured covariate matrix.}
}
\value{
A length-`N` vector.
}
\description{
Generate unmeasured confounders dependent on a measured covariate
}
