% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic.R
\name{DataFitting}
\alias{DataFitting}
\alias{DataFitting.ModelPar.ITT}
\alias{DataFitting.ModelPar.remove}
\alias{DataFitting.ModelPar.recensor}
\alias{DataFitting.ModelPar.TimeVar}
\alias{DataFitting.ModelPar.DRIV.s}
\alias{DataFitting.ModelPar.DRIV.cf.hz.ml.est}
\title{Fit a treatment-switching model}
\usage{
DataFitting(ModelPar)

\method{DataFitting}{ModelPar.ITT}(ModelPar)

\method{DataFitting}{ModelPar.remove}(ModelPar)

\method{DataFitting}{ModelPar.recensor}(ModelPar)

\method{DataFitting}{ModelPar.TimeVar}(ModelPar)

\method{DataFitting}{ModelPar.DRIV.s}(ModelPar)

\method{DataFitting}{ModelPar.DRIV.cf.hz.ml.est}(ModelPar)
}
\arguments{
\item{ModelPar}{A `ModelPar` object from [New_ModelPar()].}
}
\value{
A list with at least `Coef` and `Var`.
}
\description{
Generic that dispatches on the `ModelPar.<method>` class. Supported methods
are `ITT`, `remove`, `recensor`, `TimeVar`, `DRIV.s` and `DRIV.cf.hz.ml.est`.
}
