% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic.R
\name{DataGenerating}
\alias{DataGenerating}
\alias{DataGenerating.SimuArg.exogenous}
\alias{DataGenerating.SimuArg.endogenous}
\title{Generate simulation data}
\usage{
DataGenerating(SimuArg)

\method{DataGenerating}{SimuArg.exogenous}(SimuArg)

\method{DataGenerating}{SimuArg.endogenous}(SimuArg)
}
\arguments{
\item{SimuArg}{A `SimuArg` object from [New_SimuArg()].}
}
\value{
A list of summary rates (treatment proportion, switching, censoring).
}
\description{
Generic that dispatches on the `SimuArg.<Scenario>` class to generate (and
optionally save) replicate datasets and return summary rates.
}
